+++ examples/resources/nxos_ipv4_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ipv4_vrf.example "sys/ipv4/inst/dom-[VRF1]"