+++ examples/resources/nxos_ipv4_vrf/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_ipv4_vrf" "example" {
+  name = "VRF1"
+}